2005-05-03	Jennifer Averett <jennifer.averett@oarcorp.com>
	* libchip/network/dec21140.c, libchip/network/elnk.c,
	libchip/network/if_fxp.c, libchip/serial/mc68681.c:
	Modified parameter list to remove warnings.

+++ c/src/libchip/network/dec21140.c
@@ -925,7 +925,7 @@ dec21140_stats (struct dec21140_softc *sc)
  * Driver ioctl handler
  */
 static int
-dec21140_ioctl (struct ifnet *ifp, int command, caddr_t data)
+dec21140_ioctl (struct ifnet *ifp, u_long command, caddr_t data)
 {
    struct dec21140_softc *sc = ifp->if_softc;
    int error = 0;
